Starting work on large refactoring, first part is complete.
Changed the User class so that it doesn't make any requests with the UserApi object - I felt it was violating SRP, because if the UserApi changed, then the User class would have to change as well. The unit tests have been changed accordingly.

Added a new exception to be raised when methods which require authentication detect that it is missing. Added a method to the base API to detect it.