Updates for a few things - attempting to clean some stuff up.
It seems the update to Castle.Core 3.0 broke the interceptor for the on-the-fly authentication stuff, specifically it doesn't seem like you can intercept a protected virtual member of a base class anymore, whereas previously you could. I'm not sure if this is intentional or not as I can't find anything about it, anywhere.