Stubbed out the full api
I've stubbed out the full API for github as extension method classes in the Api.* namespace.

Each part of the namespace and the classes therin all basically map to the layout of the GithubAPI as laid out on developer.github.com.

Because they are extension methods, I've had to make a couple of changes to Github.cs to expose the RequestFactory, and to avoid exposing the internal RestSharp client, I've added Execute method overloads to it which call Execute on the internal RestSharp client.

Updated Users.cs to reflect these changes and updated the tests accordingly.